linux: stub LifeStudio:Head -- D3D9 is now the only thing left in Main
LifeStudio:Head is proprietary 2004 middleware shipped as two Win32 DLLs with
no Linux build; only the interop headers are in-tree. lsstub.cpp supplies their
five factory entry points and Init() as no-ops, the same stub pattern already
used for FMOD and Bink -- neither lsglue.cpp nor any engine file is touched.

Returning null is safe because the engine already handles a head rig failing to
build: LSHead.cpp guards every factory result, tracks success in
bMorphOk/bTexRigOk, and wraps the rig construction in try/catch. The visible
consequence is that faces do not morph -- heads render with their baked base
mesh, no FaceGen deformation and no facial animation. Nothing else is affected.
Both the stub and Main compile against the headers' own static-library mode
(LIFESTUDIOHEADAPI_EXPORTS_LIB), set PUBLIC so the declarations agree.

With LSHead.h parsing, three files that had never been reached by the compiler
turned up the usual suspects: SHMatrix row views in GView.cpp, a pointer used
as a per-unit random seed in RPGUnit.cpp, and two more dependent-base calls in
Sync.h.

Main now has exactly six errors left, all of them `D3D9.h: no such file` -- the
renderer, and nothing else.

The default (renderer-less) build still builds clean.

+++ Main/RPGUnit.cpp
@@ -454,7 +454,7 @@ void CUnit::SetHead( NDb::CComplexHead *pNewHead )
 	if ( !pNewHead )
 		return;
 	pHeadInfo = new NLSHead::CHeadInfo( pNewHead );
-	pHeadInfo->SetSeed( SRandomSeed( (int)this ) );
+	pHeadInfo->SetSeed( SRandomSeed( (int)(intptr_t)this ) );
 }
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // SetHeadInfo @0x192070 -- install an already-built live head directly (the CObj assignment open-codes the
